re render working after delete
